requirements.txt
Faker==19.3.1
kafka-python==2.0.2